Python Twitter program
Having access to the Twitter API can help you manage your social media
accounts, and allow you to mine social media for data. This can be
useful for brand promotion if you represent a business or an
organization, and it can be enjoyable and entertaining for individual
users and hobbyist programmers.

In this article, we will outline the steps necessary for you to create a
Twitter application.

We’ll then build a script in Python that uses the Tweepy library to make
use of the Twitter API.